Hide DiffGraphApplier again but provide `apply` on DiffGraph.
